Lib/reprlib.py
"""Redo the builtin repr() (representation) but przy limits on most sizes."""

__all__ = ["Repr", "repr", "recursive_repr"]

zaimportuj builtins
z itertools zaimportuj islice
spróbuj:
    z _thread zaimportuj get_ident
wyjąwszy ImportError:
    z _dummy_thread zaimportuj get_ident

def recursive_repr(fillvalue='...'):
    'Decorator to make a repr function zwróć fillvalue dla a recursive call'

    def decorating_function(user_function):
        repr_running = set()

        def wrapper(self):
            key = id(self), get_ident()
            jeżeli key w repr_running:
                zwróć fillvalue
            repr_running.add(key)
            spróbuj:
                result = user_function(self)
            w_końcu:
                repr_running.discard(key)
            zwróć result

        # Can't use functools.wraps() here because of bootstrap issues
        wrapper.__module__ = getattr(user_function, '__module__')
        wrapper.__doc__ = getattr(user_function, '__doc__')
        wrapper.__name__ = getattr(user_function, '__name__')
        wrapper.__annotations__ = getattr(user_function, '__annotations__', {})
        zwróć wrapper

    zwróć decorating_function

klasa Repr:

    def __init__(self):
        self.maxlevel = 6
        self.maxtuple = 6
        self.maxlist = 6
        self.maxarray = 5
        self.maxdict = 4
        self.maxset = 6
        self.maxfrozenset = 6
        self.maxdeque = 6
        self.maxstring = 30
        self.maxlong = 40
        self.maxother = 30

    def repr(self, x):
        zwróć self.repr1(x, self.maxlevel)

    def repr1(self, x, level):
        typename = type(x).__name__
        jeżeli ' ' w typename:
            parts = typename.split()
            typename = '_'.join(parts)
        jeżeli hasattr(self, 'repr_' + typename):
            zwróć getattr(self, 'repr_' + typename)(x, level)
        inaczej:
            zwróć self.repr_instance(x, level)

    def _repr_iterable(self, x, level, left, right, maxiter, trail=''):
        n = len(x)
        jeżeli level <= 0 oraz n:
            s = '...'
        inaczej:
            newlevel = level - 1
            repr1 = self.repr1
            pieces = [repr1(elem, newlevel) dla elem w islice(x, maxiter)]
            jeżeli n > maxiter:  pieces.append('...')
            s = ', '.join(pieces)
            jeżeli n == 1 oraz trail:  right = trail + right
        zwróć '%s%s%s' % (left, s, right)

    def repr_tuple(self, x, level):
        zwróć self._repr_iterable(x, level, '(', ')', self.maxtuple, ',')

    def repr_list(self, x, level):
        zwróć self._repr_iterable(x, level, '[', ']', self.maxlist)

    def repr_array(self, x, level):
        jeżeli nie x:
            zwróć "array('%s')" % x.typecode
        header = "array('%s', [" % x.typecode
        zwróć self._repr_iterable(x, level, header, '])', self.maxarray)

    def repr_set(self, x, level):
        jeżeli nie x:
            zwróć 'set()'
        x = _possibly_sorted(x)
        zwróć self._repr_iterable(x, level, '{', '}', self.maxset)

    def repr_frozenset(self, x, level):
        jeżeli nie x:
            zwróć 'frozenset()'
        x = _possibly_sorted(x)
        zwróć self._repr_iterable(x, level, 'frozenset({', '})',
                                   self.maxfrozenset)

    def repr_deque(self, x, level):
        zwróć self._repr_iterable(x, level, 'deque([', '])', self.maxdeque)

    def repr_dict(self, x, level):
        n = len(x)
        jeżeli n == 0: zwróć '{}'
        jeżeli level <= 0: zwróć '{...}'
        newlevel = level - 1
        repr1 = self.repr1
        pieces = []
        dla key w islice(_possibly_sorted(x), self.maxdict):
            keyrepr = repr1(key, newlevel)
            valrepr = repr1(x[key], newlevel)
            pieces.append('%s: %s' % (keyrepr, valrepr))
        jeżeli n > self.maxdict: pieces.append('...')
        s = ', '.join(pieces)
        zwróć '{%s}' % (s,)

    def repr_str(self, x, level):
        s = builtins.repr(x[:self.maxstring])
        jeżeli len(s) > self.maxstring:
            i = max(0, (self.maxstring-3)//2)
            j = max(0, self.maxstring-3-i)
            s = builtins.repr(x[:i] + x[len(x)-j:])
            s = s[:i] + '...' + s[len(s)-j:]
        zwróć s

    def repr_int(self, x, level):
        s = builtins.repr(x) # XXX Hope this isn't too slow...
        jeżeli len(s) > self.maxlong:
            i = max(0, (self.maxlong-3)//2)
            j = max(0, self.maxlong-3-i)
            s = s[:i] + '...' + s[len(s)-j:]
        zwróć s

    def repr_instance(self, x, level):
        spróbuj:
            s = builtins.repr(x)
            # Bugs w x.__repr__() can cause arbitrary
            # exceptions -- then make up something
        wyjąwszy Exception:
            zwróć '<%s instance at %#x>' % (x.__class__.__name__, id(x))
        jeżeli len(s) > self.maxother:
            i = max(0, (self.maxother-3)//2)
            j = max(0, self.maxother-3-i)
            s = s[:i] + '...' + s[len(s)-j:]
        zwróć s


def _possibly_sorted(x):
    # Since nie all sequences of items can be sorted oraz comparison
    # functions may podnieś arbitrary exceptions, zwróć an unsorted
    # sequence w that case.
    spróbuj:
        zwróć sorted(x)
    wyjąwszy Exception:
        zwróć list(x)

aRepr = Repr()
repr = aRepr.repr
